+++ crates/eips/src/eip1559/mod.rs
@@ -8,8 +8,10 @@ pub use basefee::BaseFeeParams;
 mod constants;
 pub use constants::{
     DEFAULT_BASE_FEE_MAX_CHANGE_DENOMINATOR, DEFAULT_ELASTICITY_MULTIPLIER,
-    ETHEREUM_BLOCK_GAS_LIMIT, INITIAL_BASE_FEE, MIN_PROTOCOL_BASE_FEE, MIN_PROTOCOL_BASE_FEE_U256,
+    ETHEREUM_BLOCK_GAS_LIMIT, FEE_ESTIMATION_PAST_BLOCKS, FEE_ESTIMATION_REWARD_PERCENTILE,
+    INITIAL_BASE_FEE, MIN_PROTOCOL_BASE_FEE, MIN_PROTOCOL_BASE_FEE_U256, MIN_PROTOCOL_PRIORITY_FEE,
+    MIN_PROTOCOL_PRIORITY_FEE_U256,
 };
 
 mod helpers;
-pub use helpers::calc_next_block_base_fee;
+pub use helpers::*;
